filter-repo: avoid repeatedly translating the same string with --analyze
Translating "Processed %d blob sizes" or "Processed %d commits" hundreds
of thousands or millions of times is a waste and turns out to be pretty
expensive.  Translate it once, cache the string, and then re-use it.
Note that a similar issue was noted in commit 3999349be4ed (filter-repo:
fix perf regression; avoid excessive translation, 2019-05-21), but I did
not think to check --analyze mode for similar issues back then.  Fix it
now.

Signed-off-by: Elijah Newren <newren@gmail.com>