filter-repo: make git version requirement error message more direct
Users won't know which versions of git have --mark-tags, --reencode, or
--combined-all-paths options for fast-export and diff-tree.  I didn't
either when I wrote those messages because it wasn't in a released
version of git.  Now that they are in released versions and have been
for a while, we can simplify the messages to just state which git
version is needed.

Signed-off-by: Elijah Newren <newren@gmail.com>