filter-repo: change order of commit message update
Currently, commit IDs in commit messages are translated before the
message-callback is run.  Move the commit ID translation to later so
that the message-callback can add references to commits which will be
translated.

This particularly came up in a case where a repository had been
converted from SVN to git, and there was a desire to translate SVN
revision references to the corresponding git commit, but to make sure
that the new reference was the corresponding git commit after repository
filtering rather than before repository filtering.

Signed-off-by: HassanAkbar <hassanakbar2@gmail.com>
[en: added a commit message explanation]
Signed-off-by: Elijah Newren <newren@gmail.com>