Handle tags with no date without blowing up
In some situations, git-fast-export generates tagger lines with no date,
such as:

    tag v1_7_1
    mark :758
    from :592
    original-oid 52e4c981c5ee8c4225316b418775e95dd436b6ef
    tagger Gerrit Pape <pape@smarden.org>
    data 0

This would cause git-filter-repo to fail because the regular expression
used to parse user information wouldn't match.

This commit modifies the regular expression so that it user information
without a date, and then emit a date of "0 +0000" when writing out tags
that have no date.

Addresses #413

Signed-off-by: Lars Kellogg-Stedman <lars@oddbit.com>