lint-history: fix assumption that .git/info exists
While .git/info probably existed in the past, it does not seem to be
required anymore these days.  This broke lint_non_binary_blobs, for
which we did not have a testcase.  A subsequent commit will add such
a testcase together with the `--refs` option, but first we will need
to fix that option...for now, just proactively create .git/info if
needed.

Signed-off-by: Elijah Newren <newren@gmail.com>