git-filter-repo.txt: update link to GitHub docs on purging data
GitHub updated their docs some time ago.  The third step in the
relevant section is bogus (one of the commands is filter-branch
specific despite the fact that their documentation nowadays only
references bfg-repo-cleaner and filter-repo, and the other two
commands are specific to bfg-repo-cleaner though they explicit
call it out as being relevant to filter-repo too)...but the other
parts of the page are a vast improvement over the days when it
suggested filter-branch.  Anyway, update our references to point
to the appropriate part of the updated documentation.

Signed-off-by: Elijah Newren <newren@gmail.com>