Bump python requirement to 3.6
I've wanted to use f-strings for a long time, but avoided bumping the
python version requirement for fear it'd make it a challenge to use on
some enterprise distros.  However, it has been a while, and RHEL7 is
now old (uses python 3.6.8) as is Ubuntu focal (a.k.a. 20.04; uses
python 3.8.2) and Debian buster (uses python 3.7.3).  So, even
enterprise distributions from a few generations back have a new enough
python.  Bump the requirement now.

Signed-off-by: Elijah Newren <newren@gmail.com>