t9390: skip the version test if we are in some kind of installation
The version test assumes we are in a clone of the git-filter-repo
repository, without the kind of changes that would be performed by an
installation of git-filter-repo (which may, for example, change the
shebang line or the @@LOCALEDIR@@ references).  So, if we are testing
an installation of git-filter-repo, simply skip this test.

Signed-off-by: Elijah Newren <newren@gmail.com>