filter-repo: ignore __pycache__ untracked files for fresh clone check
Some versions of python may create a bytecode compiled version of
git_filter_repo.py within the __pycache__ directory upon attempting
to run git-filter-repo, which naturally would mean we have an untracked
file and cause us to fail the fresh clone check.  Strip this file out
of the untracked files list so that we only trigger if there are other
untracked files.

Signed-off-by: Elijah Newren <newren@gmail.com>