filter-repo: allow tracking commit hashes in AncestryGraph
ID_TO_HASH and HASH_TO_ID are confusing in that it is difficult to
remember which of the four IDs (two in _orig_graph and two in _graph)
are being referenced.  And trying to use them anywhere new runs into all
kinds of problems trying to figure out what is the correct one to use.

Start tracking the original commit hashes in the AncestryGraph, where we
can document and explain how all the pieces relate.

Note that we will not use this to track hashes for RepoFilter._graph,
since the commit hashes will not be known until sometime later (due to
the _pending_renames stuff).  But we will use it to track hashes for
RepoFilter._orig_graph.

Signed-off-by: Elijah Newren <newren@gmail.com>