filter-repo: only conditionally heed the already_ran file
When users run filter-repo, we first check if they already did a
previous run.  If they recently ran filter-repo previously, then they
are probably just trying to do their history rewrite in multiple steps,
so we bypass the fresh clone check for these subsequent runs.  However,
it is possible that the user finishes their rewrite, continues as normal
for months or years, and then decides to do another rewrite.  It may be
that there is still a $GIT_DIR/filter-repo/already_ran file present from
the repository filtering from long ago, and we do not want the presence
of that file to result in the fresh clone check being bypassed.

Implement a middle ground -- if the already_ran file is older than a
day, ask the user whether to consider the current run a continuation of
an existing rewrite or if it should be considered a completely new
rewrite.  Add documentation explaining this as well.

Signed-off-by: Elijah Newren <newren@gmail.com>