git-filter-repo.txt: remove case specific explanation of commit-map file
The explanation of the commit-map file from commit f867bb6ad745
(git-filter-repo.txt: document mapping output, 2020-06-18) included an
implicit assumption that the repository filtering operation being
performed was the excise of some large file(s).  While some rewrites may
fall into that category, it is not reasonable to assume that all
rewrites are of that type.  Correct the wording.

Signed-off-by: Elijah Newren <newren@gmail.com>