filter-repo: add a new _remap_to() convenience function
Currently,
  RepoFilter._orig_graph maps old_id <-> old_hash
  RepoFilter._graph      maps new_id <-> new_hash
  _IDs                   maps old_id -> new_id
although that last map is with a twist: if commits are pruned, the
mapping is from the old_id to the nearest unpruned ancestor, if
any.

In preparation for updating the ref-map file, we need a way to map
from old_hash to new_hash (taking into account possible pruning and
mapping to unpruned ancestors).  Make a function that makes use of
these three available maps to provide such a translation layer.

Signed-off-by: Elijah Newren <newren@gmail.com>