filter-repo: add explanatory note for complicated condition
When the first parent has been pruned away, or this is a merge commit
that is up for pruning if it has no relevant file changes, then we need
to know what the file_changes are relative to the new first parent
rather than the old one (especially if the original merge was an "evil"
merge).  The conditions that are relevant can get more than a little
confusing so add a very detailed comment trying to explain the
conditions.  A later commit will actually address these conditions and
make the fixes.

Signed-off-by: Elijah Newren <newren@gmail.com>