docs: correct a couple items in explanation of ref-map file
Commit c76977213544 (filter-repo: output commit-map and ref-map in
predictable orders, 2024-10-03) changed the order of the ref-map file to
sort by refname, but the documentation still claimed that the file is in
no particular order.  Correct the oversight.

The documentation of ref-map also implied that ref-map only contained
references that were changed by the rewrite, when it actually includes
all references seen by fast-export/fast-import.  Add a clarification.

Signed-off-by: Elijah Newren <newren@gmail.com>