filter-repo: handle locally modified refs vs. automatic fetch of --sdr
As noted in the previous commit, running with the
--sensitive-data-removal flag and with locally modified refs puts users
in a pick-your-poison position.  While we provided the --no-fetch option
to allow them to choose, they might not know they need to choose.
Granted, we can't tell in all cases either, but in the common cases
where they have an origin remote and their local branch does not match
the origin remote, we can deduce that they have a locally modified
branch and then warn the user and have them choose.  This won't work if
their locally modified ref is not a branch or if they are not tracking
origin, but it will get the vast majority of cases.  And our
documentation will continue to put the responsibility on the user for
whether to run --no-fetch or not; this check is just a convenience.

Users who work on fresh clones, as we strongly suggest, don't have to
worry about any of this; things just work.

Signed-off-by: Elijah Newren <newren@gmail.com>