filter-repo: allow users to override automatic fetch forcing in --sdr
For sensitive data removals, it is important to make sure that all refs
are cleaned up, and so we attempt do a mirror-like fetch.  However, a
mirror-like fetch runs the risk of discarding local changes, and for
those folks who insist on doing their cleanup in an active repository
rather than a fresh clone, that might be detrimental.  Granted, not
fetching additional refs is also detrimental, but allow users to pick
their poison.

Signed-off-by: Elijah Newren <newren@gmail.com>