filter-repo: Accept multiline git config values
Some git config options can contain multiline values, for example
aliases. With the current simple split on = this fails with a stacktrace
so that git-filter-repo is unusable.

This patch uses a multiline regular expression to split the git config
--list string at any beginning line followed by an alphanumeric name
with '-' and '.' followed by '='. In some cases this can falsely
identify multiline values as a new variable name that does not exist.
But it should work for most usual multiline config options correctly.

Signed-off-by: Markus Schneider-Pargmann <msp@baylibre.com>