Help minimize export dumps differences when inspecting --dry-run results
Diffing the fast-export dumps with `--dry-run`, it appears git-filter-repo is using Python's default sorting algorithm for Git trees, rather than the Git-specific algorithm, which is described a bit at:
https://stackoverflow.com/questions/78981525/how-git-sort-entries-in-trees

```
$ diff -u .git/filter-repo/fast-export.original .git/filter-repo/fast-export.filtered
...
@@ -1451,25 +1329,23 @@
 D testcases/expected/case1-twenty
 D testcases/inputs/case1
 M 100755 0a13abf13c87a13ae56c1fc664baa8b44bdac1de testcases/t9390-repo-filter.sh
+M 100644 de3799fa8cd84a9ac56fad023ddf3da44eb444bb testcases/t9390/case1
 M 100644 e0c88454095bd887852791401a3b2336ad47d012 testcases/t9390/case1-filename
 M 100644 a1aa78ff11f3a40cafdcd9b531e8c1342e8a75e5 testcases/t9390/case1-ten
 M 100644 488cbd9d101b107809611940d1cd00a8f72e1b17 testcases/t9390/case1-twenty
-M 100644 de3799fa8cd84a9ac56fad023ddf3da44eb444bb testcases/t9390/case1
```

Unlike described above though, it appears the fast-export dumps treats all files as directory when sorting them (`case1` is a file in the example above), so appending "/" to all entries maintains the original tree order.

Note: git-filter-repo doesn't write tree objects directly; it has git-fast-import do that. git-fast-import states that the order of the filemodify directives given to it does not matter. So this doesn't change the resulting Git repository.