filter-repo: reduce export dump differences
While the order of files passed to `git fast-import` do not matter, we
can help users inspect --dry-run results if we can match the order that
fast-export uses.  fast-export uses a depth-first sorting order to
ensure that entries underneath a directory would appear before the
directory or some kind of file or symlink that has replaced a directory.
However, its depth-first ordering is implemented in a way that affects
things besides directories becoming symlinks or files -- in contrast to
normal lexicographic ordering where a substring would sort before a
longer string, fast-export reverses that.  This means that when we try
to inspect differences after a --dry-run, we may see a spurious
difference such as:

```
$ ./git-filter-repo --dry-run --proceed
$ diff -u .git/filter-repo/fast-export.original .git/filter-repo/fast-export.filtered
...
@@ -1451,25 +1329,23 @@
 D testcases/expected/case1-twenty
 D testcases/inputs/case1
 M 100755 0a13abf13c87a13ae56c1fc664baa8b44bdac1de testcases/t9390-repo-filter.sh
+M 100644 de3799fa8cd84a9ac56fad023ddf3da44eb444bb testcases/t9390/case1
 M 100644 e0c88454095bd887852791401a3b2336ad47d012 testcases/t9390/case1-filename
 M 100644 a1aa78ff11f3a40cafdcd9b531e8c1342e8a75e5 testcases/t9390/case1-ten
 M 100644 488cbd9d101b107809611940d1cd00a8f72e1b17 testcases/t9390/case1-twenty
-M 100644 de3799fa8cd84a9ac56fad023ddf3da44eb444bb testcases/t9390/case1
```

Fix this by modifying filter-repo to use the same sorting of paths as
fast-export when --dry-run is specified.

Signed-off-by: Sylvain Beucler <beuc@beuc.net>